Refactor host check models
