+++ epp-client/src/epp/request/host.rs
@@ -1,6 +1,5 @@
 //! Types for EPP host requests
 
-pub mod check;
 pub mod create;
 pub mod delete;
 pub mod info;
