+++ epp-client/src/epp/response/host.rs
@@ -1,6 +1,5 @@
 //! Types for EPP host responses
 
-pub mod check;
 pub mod create;
 pub mod delete;
 pub mod info;
