+++ epp-client/src/lib.rs
@@ -106,6 +106,7 @@ pub mod contact;
 pub mod domain;
 pub mod epp;
 pub mod error;
+pub mod host;
 pub use connection::client::EppClient;
 
 #[cfg(test)]
