+++ epp-client/src/tests/se.rs
@@ -30,6 +30,7 @@ mod request {
     use crate::epp::request::{EppHello, EppLogin, EppLogout};
     use crate::epp::xml::EppXml;
     use crate::epp::*;
+    use crate::host::check::EppHostCheck;
     use chrono::{DateTime, NaiveDate};
     use std::str::FromStr;
